Split V1 spec truth tests
