+++ tests/docs_truth/v1_spec.rs
@@ -1,477 +1,8 @@
-use super::*;
-
-#[test]
-fn v1_spec_records_phase_one_feature_gates_and_test_backlog() {
-    let spec = read("docs/V1_SPEC.md");
-
-    for required in [
-        "Status: v1 draft",
-        "Feature Matrix",
-        "implemented",
-        "gated",
-        "experimental",
-        "removed",
-        "Sync/Async effects",
-        "async_scheduler_intrinsics_are_rejected_as_gated_not_unknown",
-        "stdlib_async_runtime_import_is_gated_before_loading_sketch",
-        "module_graph_gates_stdlib_async_runtime_import_before_loading_sketch",
-        "emit_json_diagnostics_async_runtime_import_gate_schema_matches_golden",
-        "stdlib_sync_runtime_import_is_gated_before_loading_sketch",
-        "module_graph_gates_stdlib_sync_runtime_import_before_loading_sketch",
-        "emit_json_diagnostics_sync_runtime_import_gate_schema_matches_golden",
-        "async task enqueue",
-        "async yield",
-        "atomic_intrinsics_are_rejected_as_effect_gates",
-        "@builtin.atomic_load",
-        "@builtin.atomic_store",
-        "@builtin.atomic_add",
-        "@builtin.atomic_sub",
-        "@builtin.atomic_cas",
-        "@builtin.atomic_xchg",
-        "@builtin.fence",
-        "Typed allocators",
-        "raw_memory_intrinsics_are_rejected_as_allocator_gates",
-        "byte_memory_intrinsics_are_rejected_as_allocator_gates",
-        "sync_and_async_typed_allocator_modes_are_rejected_as_gated_not_unknown",
-        "stdlib_allocator_import_is_gated_before_loading_sketch",
-        "module_graph_gates_stdlib_allocator_import_before_loading_sketch",
-        "emit_json_diagnostics_allocator_import_gate_schema_matches_golden",
-        "@builtin.raw_allocate",
-        "@builtin.raw_deallocate",
-        "@builtin.raw_reallocate",
-        "@builtin.memcpy",
-        "@builtin.memmove",
-        "@builtin.memset",
-        "@builtin.memcmp",
-        "Type matching",
-        "Behavior association",
-        "AST traversal",
-        "`Channel` remains an experimental stdlib channel",
-        "semantic_status: \"unchecked\"",
-        "typed JSON is explicitly marked checked",
-        "diagnostics JSON is explicitly",
-        "docs/DIAGNOSTICS.md",
-        "JSON-stable public diagnostic codes",
-        "broader diagnostic-code coverage is still required",
-        "context.kind = \"feature_gate\"",
-        "Developer UX and Agent UX",
-        "product requirements, not polish",
-        "MoonBit-style toolchain integration",
-        "VS Code extension remains a constrained editor wrapper",
-        "`zen lsp` remains gated",
-        "Agent-readable diagnostics",
-        "machine-readable project graph",
-        "structured fix suggestions",
-        "quiet deterministic commands",
-        "must not advertise unsupported language-server binaries",
-        "Type.implements(Behavior)",
-        "non-generic explicit behavior associations",
-        "Actors in std",
-        "bare_actor_framework_types_are_rejected_as_gated_not_unknown",
-        "stdlib_actor_framework_import_is_gated_before_loading_sketch",
-        "module_graph_gates_stdlib_actor_framework_import_before_loading_sketch",
-        "emit_json_diagnostics_actor_import_gate_schema_matches_golden",
-        "JSON/YAML IR boundaries",
-        "comptime_type_match_intrinsic_is_rejected_as_gated_not_unknown",
-        "primitive_and_enum_type_match_intrinsics_are_rejected_as_gated_not_unknown",
-        "comptime type matching",
-        "gated until typed metadata and derive lowering exist",
-        "Ownership and raw pointer operations",
-        "raw_pointer_intrinsics_are_rejected_as_ownership_gates",
-        "@builtin.gep",
-        "@builtin.gep_struct",
-        "@builtin.raw_ptr_cast",
-        "@builtin.ptr_to_int",
-        "@builtin.int_to_ptr",
-        "@builtin.load<T>",
-        "@builtin.store<T>",
-        "Host syscalls",
-        "syscall_intrinsics_are_rejected_as_host_effect_gates",
-        "@builtin.syscall0",
-        "@builtin.syscall6",
-        "zen emit-json ast <file>",
-        "zen emit-json symbols <file>",
-        "zen emit-json typed <file>",
-        "zen emit-json diagnostics <file>",
-        "semantic acceptance must use typed",
-        "emit_json_ast_rejects_hand_authored_json_before_unchecked_ir_override",
-        "emit_json_symbols_rejects_hand_authored_json_before_resolver_override",
-        "emit_json_ast_module_graph_schema_matches_golden",
-        "emit_json_symbols_module_graph_schema_matches_golden",
-        "emit_json_symbols_generic_method_schema_matches_golden",
-        "emit_json_symbols_generic_type_impl_methods_schema_matches_golden",
-        "emit_json_symbols_generic_self_method_schema_matches_golden",
-        "emit_json_symbols_generic_method_worklist_schema_matches_golden",
-        "emit_json_symbols_generic_ufc_function_schema_matches_golden",
-        "emit_json_symbols_generic_option_schema_matches_golden",
-        "emit_json_symbols_generic_result_schema_matches_golden",
-        "emit_json_symbols_generic_result_method_schema_matches_golden",
-        "emit_json_symbols_generic_behavior_association_schema_matches_golden",
-        "emit_json_symbols_generic_behavior_bound_ufcs_schema_matches_golden",
-        "Box.get<T>",
-        "Box.replace<T>",
-        "Box<T>.impl",
-        "Box.copy<T>",
-        "Option.copy<T>",
-        "inner<T>",
-        "id<T>",
-        "12.id<i32>()",
-        "id_i32",
-        "Box.get_inner<T>",
-        "Box.get_inner_i32",
-        "inner_i32",
-        "Option<T>",
-        "unwrap_or<T>",
-        "Result<T, E>",
-        "unwrap_or<T, E>",
-        "Result.unwrap_or<T, E>",
-        "self: Self",
-        "Json<StaticString>",
-        "Point.encode__Json_Point",
-        "Json<Point>",
-        "emit_json_diagnostics_rejects_hand_authored_json_before_diagnostic_override",
-        "emit_json_diagnostics_removed_return_schema_matches_golden",
-        "emit_json_diagnostics_behavior_derive_gate_schema_matches_golden",
-        "emit_json_diagnostics_generic_association_gate_schema_matches_golden",
-        "emit_json_diagnostics_typed_allocator_effect_gate_schema_matches_golden",
-        "emit_json_diagnostics_sync_effect_gate_schema_matches_golden",
-        "emit_json_diagnostics_async_effect_gate_schema_matches_golden",
-        "emit_json_diagnostics_dynamic_string_gate_schema_matches_golden",
-        "emit_json_diagnostics_generic_dynamic_string_gate_schema_matches_golden",
-        "emit_json_diagnostics_actor_type_gate_schema_matches_golden",
-        "emit_json_diagnostics_bare_actor_type_gate_schema_matches_golden",
-        "emit_json_diagnostics_type_match_gate_schema_matches_golden",
-        "emit_json_diagnostics_async_intrinsic_gate_schema_matches_golden",
-        "emit_json_diagnostics_atomic_gate_schema_matches_golden",
-        "emit_json_diagnostics_syscall_gate_schema_matches_golden",
-        "emit_json_diagnostics_raw_allocate_gate_schema_matches_golden",
-        "emit_json_diagnostics_byte_memory_gate_schema_matches_golden",
-        "emit_json_diagnostics_raw_pointer_gate_schema_matches_golden",
-        "emit_json_diagnostics_range_gate_schema_matches_golden",
-        "emit_json_diagnostics_raise_gate_schema_matches_golden",
-        "emit_json_diagnostics_await_gate_schema_matches_golden",
-        "emit_json_diagnostics_generic_function_arity_schema_matches_golden",
-        "emit_json_diagnostics_nongeneric_function_type_args_schema_matches_golden",
-        "emit_json_diagnostics_nongeneric_module_function_type_args_schema_matches_golden",
-        "emit_json_diagnostics_nongeneric_builtin_function_type_args_schema_matches_golden",
-        "emit_json_diagnostics_generic_function_type_arg_annotation_arity_schema_matches_golden",
-        "emit_json_diagnostics_generic_method_type_arg_annotation_arity_schema_matches_golden",
-        "emit_json_diagnostics_nongeneric_method_type_args_schema_matches_golden",
-        "emit_json_diagnostics_generic_method_type_arg_annotation_missing_args_schema_matches_golden",
-        "emit_json_diagnostics_closure_param_annotation_type_arg_arity_schema_matches_golden",
-        "emit_json_diagnostics_closure_return_annotation_missing_args_schema_matches_golden",
-        "emit_json_diagnostics_cast_target_annotation_type_arg_arity_schema_matches_golden",
-        "emit_json_diagnostics_cast_target_annotation_missing_args_schema_matches_golden",
-        "emit_json_diagnostics_nested_generic_annotation_inner_arity_schema_matches_golden",
-        "emit_json_diagnostics_nested_generic_instantiation_inner_arity_schema_matches_golden",
-        "emit_json_diagnostics_function_type_parameter_annotation_arity_schema_matches_golden",
-        "emit_json_diagnostics_function_type_return_annotation_missing_args_schema_matches_golden",
-        "emit_json_diagnostics_pointer_inner_generic_annotation_arity_schema_matches_golden",
-        "emit_json_diagnostics_slice_inner_generic_annotation_missing_args_schema_matches_golden",
-        "emit_json_diagnostics_array_inner_generic_annotation_arity_schema_matches_golden",
-        "emit_json_diagnostics_generic_struct_local_annotation_arity_schema_matches_golden",
-        "emit_json_diagnostics_generic_struct_local_annotation_missing_args_schema_matches_golden",
-        "emit_json_diagnostics_generic_enum_local_annotation_arity_schema_matches_golden",
-        "emit_json_diagnostics_generic_enum_local_annotation_missing_args_schema_matches_golden",
-        "emit_json_diagnostics_generic_struct_constructor_arity_schema_matches_golden",
-        "emit_json_diagnostics_generic_struct_constructor_missing_args_schema_matches_golden",
-        "emit_json_diagnostics_nongeneric_struct_constructor_type_args_schema_matches_golden",
-        "emit_json_diagnostics_generic_enum_constructor_arity_schema_matches_golden",
-        "emit_json_diagnostics_generic_enum_constructor_missing_args_schema_matches_golden",
-        "emit_json_diagnostics_nongeneric_enum_constructor_type_args_schema_matches_golden",
-        "emit_json_diagnostics_generic_struct_annotation_arity_schema_matches_golden",
-        "emit_json_diagnostics_nongeneric_struct_annotation_type_args_schema_matches_golden",
-        "emit_json_diagnostics_nongeneric_enum_annotation_type_args_schema_matches_golden",
-        "emit_json_diagnostics_generic_enum_annotation_arity_schema_matches_golden",
-        "emit_json_diagnostics_generic_struct_annotation_missing_args_schema_matches_golden",
-        "emit_json_diagnostics_generic_enum_annotation_missing_args_schema_matches_golden",
-        "emit_json_diagnostics_generic_result_method_arity_schema_matches_golden",
-        "emit_json_diagnostics_generic_result_method_bound_schema_matches_golden",
-        "emit_json_diagnostics_generic_result_method_inference_schema_matches_golden",
-        "emit_json_diagnostics_generic_behavior_overlap_schema_matches_golden",
-        "emit_json_typed_rejects_hand_authored_json_before_checked_ir_override",
-        "emit_json_typed_generic_method_schema_matches_golden",
-        "emit_json_typed_generic_type_impl_methods_schema_matches_golden",
-        "emit_json_typed_generic_self_method_schema_matches_golden",
-        "emit_json_typed_generic_method_worklist_schema_matches_golden",
-        "emit_json_typed_generic_ufc_function_schema_matches_golden",
-        "emit_json_typed_generic_option_schema_matches_golden",
-        "emit_json_typed_generic_result_schema_matches_golden",
-        "emit_json_typed_generic_result_method_schema_matches_golden",
-        "emit_json_typed_nested_generic_result_schema_matches_golden",
-        "emit_json_typed_generic_behavior_association_schema_matches_golden",
-        "emit_json_typed_generic_behavior_bound_ufcs_schema_matches_golden",
-        "Point.encode__Json_Point",
-        "emit_json_build_graph_rejects_hand_authored_json_before_graph_override",
-        "schema_version: 0",
-        "emit_json_hir_outputs_checked_declaration_graph",
-        "emit_json_hir_outputs_enum_function_and_global_declarations",
-        "emit_json_hir_declaration_schema_matches_golden",
-        "emit_json_hir_generic_method_worklist_schema_matches_golden",
-        "emit_json_hir_generic_ufc_function_schema_matches_golden",
-        "emit_json_hir_generic_option_schema_matches_golden",
-        "emit_json_hir_generic_result_schema_matches_golden",
-        "emit_json_hir_nested_generic_result_schema_matches_golden",
-        "emit_json_hir_generic_behavior_association_schema_matches_golden",
-        "emit_json_hir_generic_behavior_bound_ufcs_schema_matches_golden",
-        "emit_json_hir_rejects_hand_authored_json_before_ir_override",
-        "emit_json_mir_outputs_checked_minimal_function_graph",
-        "emit_json_mir_minimal_function_schema_matches_golden",
-        "emit_json_mir_outputs_match_arm_schema",
-        "emit_json_mir_match_schema_matches_golden",
-        "emit_json_mir_generic_method_worklist_schema_matches_golden",
-        "emit_json_mir_generic_ufc_function_schema_matches_golden",
-        "emit_json_mir_generic_option_schema_matches_golden",
-        "emit_json_mir_generic_result_schema_matches_golden",
-        "emit_json_mir_nested_generic_result_schema_matches_golden",
-        "emit_json_mir_generic_behavior_association_schema_matches_golden",
-        "emit_json_mir_generic_behavior_bound_ufcs_schema_matches_golden",
-        "emit_json_mir_rejects_hand_authored_json_before_ir_override",
-        "emit_json_layout_outputs_checked_type_layouts",
-        "emit_json_layout_basic_schema_matches_golden",
-        "emit_json_layout_outputs_compound_type_layout_entries",
-        "emit_json_layout_compound_schema_matches_golden",
-        "emit_json_layout_generic_option_schema_matches_golden",
-        "emit_json_layout_generic_result_schema_matches_golden",
-        "emit_json_layout_nested_generic_result_schema_matches_golden",
-        "emit_json_layout_rejects_hand_authored_json_before_layout_override",
-        "emit_json_target_yaml_validates_minimal_target_schema",
-        "emit_json_target_yaml_validates_backend_schema",
-        "emit_json_target_yaml_validates_c_backend_flags",
-        "emit_json_target_yaml_backend_schema_matches_golden",
-        "emit_json_target_yaml_rejects_empty_c_backend_flags",
-        "emit_json_target_yaml_rejects_layout_overrides",
-        "emit_json_target_yaml_rejects_unsupported_backend_codegen",
-        "emit_json_build_graph_project_schema_matches_golden",
-        "emit_json_build_graph_host_effect_schema_matches_golden",
-        "emit_json_build_graph_target_metadata_schema_matches_golden",
-        "build.zen",
-        "zen test build.zen",
-        "test target execution",
-        "Executable target dependencies compile",
-        "dependency cycles",
-        "dependency-ordered multi-executable build tests",
-        "Deterministic build graph compiles executable and test targets",
-        "Accepted Syntax Forms",
-        "Test Evidence",
-        "StaticString",
-        "baked into the program",
-        "String literals as baked `StaticString`",
-        "interpolation as non-owning `StaticString` views",
-        "only literal text is guaranteed",
-        "baked program storage",
-        "interpolation must not imply allocator-backed",
-        "allocator-backed `String`",
-        "dynamic_string_type_is_rejected_as_allocator_backed_gate",
-        "Source-level `String` use currently reports a gated",
-        "Planned Positive Test",
-        "Planned Negative Test",
-        "resolver_records_value_symbol_generic_parameter_counts",
-        "resolver_records_value_symbol_function_type_metadata",
-        "resolver_records_value_symbol_generic_bounds",
-        "resolver_records_type_and_behavior_generic_parameter_counts",
-        "resolver_rejects_duplicate_type_parameter_names",
-        "resolver_records_generic_struct_field_types",
-        "resolver_records_generic_enum_variant_payload_types",
-        "resolver_records_generic_behavior_method_signatures",
-        "resolver_records_generic_behavior_function_type_method_signatures",
-        "parse_public_behavior_declaration",
-        "resolver_records_public_visibility_for_exported_declarations",
-        "resolver_rejects_duplicate_signature_parameter_names",
-        "check_program_with_symbols_validates_resolver_function_type_parameter_metadata",
-        "check_program_with_symbols_validates_resolver_function_type_return_metadata",
-        "check_program_with_symbols_validates_resolver_function_typed_signature_metadata",
-        "collect_declarations_with_symbols_uses_resolver_function_type_metadata",
-        "collect_declarations_with_symbols_uses_resolver_generic_function_template_metadata",
-        "collect_declarations_with_symbols_uses_resolver_generic_method_template_metadata",
-        "check_program_with_symbols_validates_resolver_function_type_parameter_names",
-        "check_program_with_symbols_validates_resolver_function_type_parameter_bound_refs",
-        "check_program_with_symbols_validates_resolver_type_parameter_names",
-        "behavior_impl_generic_behavior_type_arg_bound_passes_when_satisfied",
-        "behavior_impl_generic_behavior_type_arg_bound_failure_is_error",
-        "behavior_extends_generic_parent_accepts_child_type_parameter_arg",
-        "check_program_with_symbols_validates_resolver_generic_struct_field_types",
-        "check_program_with_symbols_validates_resolver_generic_enum_payload_types",
-        "check_program_with_symbols_validates_resolver_generic_behavior_method_signatures",
-        "check_program_with_symbols_validates_resolver_generic_behavior_function_type_method_signatures",
-        "check_program_with_symbols_validates_resolver_behavior_type_parameter_bounds",
-        "resolver_rejects_duplicate_struct_field_names",
-        "resolver_rejects_duplicate_struct_literal_fields",
-        "resolver_rejects_unknown_struct_literal_fields",
-        "resolver_rejects_missing_struct_literal_fields",
-        "resolver_rejects_unknown_struct_literal_types",
-        "resolver_allows_same_variant_names_in_different_enums",
-        "resolver_rejects_duplicate_variant_names_in_same_enum",
-        "resolver_rejects_unknown_enum_variant_expressions",
-        "resolver_rejects_missing_enum_variant_payload_expressions",
-        "resolver_rejects_unexpected_enum_variant_payload_expressions",
-        "tests/zen/duplicate_enum_variant_names.zen",
-        "check_module_graph_entry_seeds_imported_function_type_signatures",
-        "check_module_graph_entry_specializes_imported_generic_functions",
-        "check_module_graph_entry_specializes_imported_generic_enums",
-        "check_module_graph_entry_seeds_public_methods_for_imported_types",
-        "check_module_graph_entry_specializes_public_generic_methods_for_imported_types",
-        "tests/zen/multi_file_generic/main.zen",
-        "tests/zen/multi_file_generic_imported_type_dependency/main.zen",
-        "tests/zen/multi_file_generic_imported_transitive_dependency/main.zen",
-        "tests/zen/multi_file_type_method_worklist/main.zen",
-        "tests/zen/multi_file_type_method_method_dependency/main.zen",
-        "tests/zen/multi_file_type_method_imported_dependency/main.zen",
-        "tests/zen/multi_file_type_impl_imported_type_dependency/main.zen",
-        "tests/zen/multi_file_behavior_bound/main.zen",
-        "tests/zen/multi_file_behavior_inheritance/main.zen",
-        "tests/zen/multi_file_imported_behavior_impl/main.zen",
-        "tests/zen/multi_file_imported_behavior_default/main.zen",
-        "tests/zen/multi_file_imported_impl_imported_behavior/main.zen",
-        "tests/zen/multi_file_imported_child_parent_dispatch/main.zen",
-        "tests/zen/multi_file_imported_behavior_requires/main.zen",
-        "tests/zen/multi_file_imported_function_imported_behavior_bound/main.zen",
-        "tests/zen/multi_file_imported_function_param_type_dependency/main.zen",
-        "tests/zen/multi_file_imported_function_return_type_dependency/main.zen",
-        "tests/zen/multi_file_imported_function_imported_return_type_behavior/main.zen",
-        "tests/zen/multi_file_imported_generic_function_return_enum_dependency/main.zen",
-        "Generic specialization for functions, structs, enums, and methods",
-        "generic_specializations_emit_each_generated_c_definition_once",
-        "generic_specializations::enum_generated_c::enum_specializations_do_not_emit_unspecialized_c_symbols",
-        "generic_specializations::method_worklist_generated_c::method_and_worklist_specializations_do_not_emit_unspecialized_c_symbols",
-        "generic_specializations::multifile_generated_c::enum_dependencies::multi_file_generic_enum_specializations_do_not_emit_unspecialized_c_symbols",
-        "generic_specializations::multifile_generated_c::method_worklist_dependencies::multi_file_generic_method_and_worklist_specializations_do_not_emit_unspecialized_c_symbols",
-        "Explicit behavior association proving ground",
-        "tests/zen/behavior_json_explicit_impl.zen",
-        "tests/zen/behavior_json_generic_association.zen",
-        "tests/zen/behavior_distinct_generic_specialization_dispatch.zen",
-        "tests/zen/behavior_json_generic_bound_ufcs.zen",
-        "tests/zen/multi_file_imported_behavior_requires/main.zen",
-        "tests/zen/multi_file_behavior_inheritance/main.zen",
-        "generic_diagnostics::behavior_impl_for_unspecialized_generic_type_is_error",
-        "generic_diagnostics::generic_behavior_bound_unknown_method_is_error",
-        "generated_c_call_definition_scan_reports_missing_generated_calls",
-        "generated_c_definition_count_ignores_prototypes",
-        "imported_function_signature_type_dependencies_are_not_directly_visible",
-        "imported_behavior_extends_requires_parent_methods",
-        "imported_behavior_extends_imported_parent_requires_parent_methods",
-        "imported_behavior_extends_requires_transitive_parent_methods",
-        "resolver_rejects_duplicate_behavior_impl_edges",
-        "resolver_rejects_duplicate_behavior_required_edges",
-        "resolver_rejects_duplicate_behavior_parent_edges",
-        "tests/zen/generic_method_worklist.zen",
-        "resolver_records_behavior_impl_methods_as_value_symbols",
-        "resolver_records_behavior_impl_function_type_methods",
-        "check_program_with_symbols_validates_resolver_impl_method_signature",
-        "check_program_with_symbols_validates_resolver_impl_function_type_signature",
-        "imported_private_behavior_impl_methods_are_not_directly_visible",
-        "resolver_rejects_duplicate_behavior_method_names",
-        "check_program_with_symbols_validates_resolver_generic_behavior_impl_names",
-        "check_program_with_symbols_validates_resolver_generic_behavior_required_names",
-        "check_program_with_symbols_validates_resolver_generic_behavior_parent_names",
-        "check_program_with_symbols_validates_resolver_generic_behavior_impl_refs",
-        "check_program_with_symbols_validates_resolver_generic_behavior_required_refs",
-        "check_program_with_symbols_validates_resolver_generic_behavior_parent_refs",
-        "collect_declarations_with_symbols_uses_resolver_behavior_impl_metadata",
-        "collect_declarations_with_symbols_uses_resolver_behavior_parent_metadata",
-        "resolver_records_method_signatures_as_value_symbols",
-        "resolver_records_method_function_type_signatures",
-        "check_program_with_symbols_validates_resolver_method_signature",
-        "check_program_with_symbols_validates_resolver_method_function_type_signature",
-        "check_module_graph_entry_does_not_seed_private_methods_for_imported_types",
-        "tests/zen/multi_file_type_method/main.zen",
-        "test_multi_file_type_method_worklist_imports",
-        "test_multi_file_type_method_method_dependency_imports",
-        "test_multi_file_type_method_imported_dependency_imports",
-        "imported_type_method_worklist_helpers_are_not_directly_visible",
-        "imported_type_method_dependencies_are_not_directly_visible",
-        "imported_type_method_imported_dependencies_are_not_directly_visible",
-        "imported_type_impl_imported_type_dependencies_are_not_directly_visible",
-        "imported_generic_function_imported_type_dependencies_are_not_directly_visible",
-        "imported_generic_function_transitive_dependencies_are_not_directly_visible",
-        "parse_impl_block",
-        "parse_generic_impl_block_hoists_receiver_type_params_to_methods",
-        "resolver_accepts_non_behavior_impl_blocks_as_method_symbols",
-        "resolver_rejects_duplicate_non_behavior_impl_method_names",
-        "resolver_rejects_non_behavior_impl_method_colliding_with_top_level_method",
-        "tests/zen/type_impl_methods.zen",
-        "tests/zen/generic_type_impl_methods.zen",
-        "tests/zen/multi_file_type_impl/main.zen",
-        "test_generic_type_impl_methods",
-        "imported_private_type_impl_methods_are_not_visible",
-        "resolver_records_struct_function_type_fields",
-        "check_program_with_symbols_validates_resolver_struct_function_type_fields",
-        "check_program_with_symbols_validates_resolver_struct_typed_field_metadata",
-        "collect_declarations_with_symbols_uses_resolver_struct_field_metadata",
-        "resolver_records_enum_function_type_payloads",
-        "check_program_with_symbols_validates_resolver_enum_function_type_payloads",
-        "check_program_with_symbols_validates_resolver_enum_typed_payload_metadata",
-        "collect_declarations_with_symbols_uses_resolver_enum_payload_metadata",
-        "resolver_records_generic_enum_function_type_payloads",
-        "check_program_with_symbols_validates_resolver_generic_enum_function_type_payloads",
-        "resolver_records_behavior_function_type_method_signatures",
-        "check_program_with_symbols_validates_resolver_behavior_function_type_method_signatures",
-        "check_program_with_symbols_validates_resolver_behavior_method_types",
-        "collect_declarations_with_symbols_uses_resolver_behavior_method_metadata",
-        "resolver_records_generic_behavior_function_type_method_signatures",
-        "check_program_with_symbols_validates_resolver_generic_behavior_function_type_method_signatures",
-        "resolver_records_behavior_impl_method_body_locals",
-        "check_program_with_symbols_requires_resolver_impl_method_body_locals",
-        "resolver_records_behavior_default_method_body_locals",
-        "check_program_with_symbols_requires_resolver_behavior_default_locals",
-        "resolver_records_struct_field_default_locals",
-        "check_program_with_symbols_requires_resolver_struct_field_default_locals",
-        "resolver_records_top_level_expr_locals",
-        "check_program_with_symbols_requires_resolver_top_level_expr_locals",
-        "resolver_records_closure_locals",
-        "check_program_with_symbols_requires_resolver_closure_locals",
-        "resolver_records_mutable_closure_parameter_locals",
-        "check_program_with_symbols_validates_resolver_closure_parameter_mutability",
-        "resolver_records_pattern_locals",
-        "check_program_with_symbols_requires_resolver_pattern_locals",
-        "resolver_records_same_name_locals_in_distinct_scopes",
-        "check_program_with_symbols_validates_resolver_local_mutability_by_scope",
-    ] {
-        assert!(
-            spec.contains(required),
-            "docs/V1_SPEC.md is missing Phase 1 requirement: {required}"
-        );
-    }
-
-    assert!(
-        !spec.contains("Deterministic build graph creates one executable target"),
-        "docs/V1_SPEC.md still describes the build.zen backlog as single-executable only"
-    );
-
-    let backlog = spec
-        .split("## Required Test Backlog")
-        .nth(1)
-        .expect("V1 spec should contain required test backlog");
-    assert!(
-        !spec.contains("String literals and interpolation as `StaticString`"),
-        "docs/V1_SPEC.md should not imply interpolation is baked program storage"
-    );
-    assert!(
-        !backlog.contains("| `build.zen` |"),
-        "docs/V1_SPEC.md should not list constrained build.zen execution as only planned backlog"
-    );
-    assert!(
-        !backlog.contains("| AST traversal |"),
-        "docs/V1_SPEC.md should not list AST traversal in the no-minimum-proof backlog after AST JSON boundary tests exist"
-    );
-    assert!(
-        !backlog.contains("| Behavior association |"),
-        "docs/V1_SPEC.md should not treat explicit behavior association proving-ground coverage as unproven backlog"
-    );
-    assert!(
-        backlog.contains("| Generated/fallback behavior association |"),
-        "docs/V1_SPEC.md should keep generated/fallback behavior association in the required backlog"
-    );
-    assert!(
-        !spec.contains("| Strict resolver, symbol IDs, privacy | gated |"),
-        "docs/V1_SPEC.md should not describe implemented resolver/module privacy evidence as gated"
-    );
-    assert!(
-        !spec.contains("`ActorRef`, `Mailbox`, `Channel`, and `Supervisor`"),
-        "docs/V1_SPEC.md should not imply Channel is a globally gated actor builtin"
-    );
-    assert!(
-        !spec.contains("sync and async allocator tests still required"),
-        "docs/V1_SPEC.md should not keep stale allocator-test backlog wording after Sync/Async allocator gate coverage exists"
-    );
-}
+#[path = "v1_spec/build_graph_and_strings.rs"]
+mod build_graph_and_strings;
+#[path = "v1_spec/feature_matrix.rs"]
+mod feature_matrix;
+#[path = "v1_spec/implementation_evidence.rs"]
+mod implementation_evidence;
+#[path = "v1_spec/json_boundaries.rs"]
+mod json_boundaries;
