+++ tests/docs_truth/v1_spec/json_boundaries.rs
@@ -1,0 +1,170 @@
+use super::super::*;
+
+#[test]
+fn v1_spec_records_json_boundaries_and_golden_coverage() {
+    let spec = read("docs/V1_SPEC.md");
+
+    for required in [
+        "zen emit-json ast <file>",
+        "zen emit-json symbols <file>",
+        "zen emit-json typed <file>",
+        "zen emit-json diagnostics <file>",
+        "semantic acceptance must use typed",
+        "emit_json_ast_rejects_hand_authored_json_before_unchecked_ir_override",
+        "emit_json_symbols_rejects_hand_authored_json_before_resolver_override",
+        "emit_json_ast_module_graph_schema_matches_golden",
+        "emit_json_symbols_module_graph_schema_matches_golden",
+        "emit_json_symbols_generic_method_schema_matches_golden",
+        "emit_json_symbols_generic_type_impl_methods_schema_matches_golden",
+        "emit_json_symbols_generic_self_method_schema_matches_golden",
+        "emit_json_symbols_generic_method_worklist_schema_matches_golden",
+        "emit_json_symbols_generic_ufc_function_schema_matches_golden",
+        "emit_json_symbols_generic_option_schema_matches_golden",
+        "emit_json_symbols_generic_result_schema_matches_golden",
+        "emit_json_symbols_generic_result_method_schema_matches_golden",
+        "emit_json_symbols_generic_behavior_association_schema_matches_golden",
+        "emit_json_symbols_generic_behavior_bound_ufcs_schema_matches_golden",
+        "Box.get<T>",
+        "Box.replace<T>",
+        "Box<T>.impl",
+        "Box.copy<T>",
+        "Option.copy<T>",
+        "inner<T>",
+        "id<T>",
+        "12.id<i32>()",
+        "id_i32",
+        "Box.get_inner<T>",
+        "Box.get_inner_i32",
+        "inner_i32",
+        "Option<T>",
+        "unwrap_or<T>",
+        "Result<T, E>",
+        "unwrap_or<T, E>",
+        "Result.unwrap_or<T, E>",
+        "self: Self",
+        "Json<StaticString>",
+        "Point.encode__Json_Point",
+        "Json<Point>",
+        "emit_json_diagnostics_rejects_hand_authored_json_before_diagnostic_override",
+        "emit_json_diagnostics_removed_return_schema_matches_golden",
+        "emit_json_diagnostics_behavior_derive_gate_schema_matches_golden",
+        "emit_json_diagnostics_generic_association_gate_schema_matches_golden",
+        "emit_json_diagnostics_typed_allocator_effect_gate_schema_matches_golden",
+        "emit_json_diagnostics_sync_effect_gate_schema_matches_golden",
+        "emit_json_diagnostics_async_effect_gate_schema_matches_golden",
+        "emit_json_diagnostics_dynamic_string_gate_schema_matches_golden",
+        "emit_json_diagnostics_generic_dynamic_string_gate_schema_matches_golden",
+        "emit_json_diagnostics_actor_type_gate_schema_matches_golden",
+        "emit_json_diagnostics_bare_actor_type_gate_schema_matches_golden",
+        "emit_json_diagnostics_type_match_gate_schema_matches_golden",
+        "emit_json_diagnostics_async_intrinsic_gate_schema_matches_golden",
+        "emit_json_diagnostics_atomic_gate_schema_matches_golden",
+        "emit_json_diagnostics_syscall_gate_schema_matches_golden",
+        "emit_json_diagnostics_raw_allocate_gate_schema_matches_golden",
+        "emit_json_diagnostics_byte_memory_gate_schema_matches_golden",
+        "emit_json_diagnostics_raw_pointer_gate_schema_matches_golden",
+        "emit_json_diagnostics_range_gate_schema_matches_golden",
+        "emit_json_diagnostics_raise_gate_schema_matches_golden",
+        "emit_json_diagnostics_await_gate_schema_matches_golden",
+        "emit_json_diagnostics_generic_function_arity_schema_matches_golden",
+        "emit_json_diagnostics_nongeneric_function_type_args_schema_matches_golden",
+        "emit_json_diagnostics_nongeneric_module_function_type_args_schema_matches_golden",
+        "emit_json_diagnostics_nongeneric_builtin_function_type_args_schema_matches_golden",
+        "emit_json_diagnostics_generic_function_type_arg_annotation_arity_schema_matches_golden",
+        "emit_json_diagnostics_generic_method_type_arg_annotation_arity_schema_matches_golden",
+        "emit_json_diagnostics_nongeneric_method_type_args_schema_matches_golden",
+        "emit_json_diagnostics_generic_method_type_arg_annotation_missing_args_schema_matches_golden",
+        "emit_json_diagnostics_closure_param_annotation_type_arg_arity_schema_matches_golden",
+        "emit_json_diagnostics_closure_return_annotation_missing_args_schema_matches_golden",
+        "emit_json_diagnostics_cast_target_annotation_type_arg_arity_schema_matches_golden",
+        "emit_json_diagnostics_cast_target_annotation_missing_args_schema_matches_golden",
+        "emit_json_diagnostics_nested_generic_annotation_inner_arity_schema_matches_golden",
+        "emit_json_diagnostics_nested_generic_instantiation_inner_arity_schema_matches_golden",
+        "emit_json_diagnostics_function_type_parameter_annotation_arity_schema_matches_golden",
+        "emit_json_diagnostics_function_type_return_annotation_missing_args_schema_matches_golden",
+        "emit_json_diagnostics_pointer_inner_generic_annotation_arity_schema_matches_golden",
+        "emit_json_diagnostics_slice_inner_generic_annotation_missing_args_schema_matches_golden",
+        "emit_json_diagnostics_array_inner_generic_annotation_arity_schema_matches_golden",
+        "emit_json_diagnostics_generic_struct_local_annotation_arity_schema_matches_golden",
+        "emit_json_diagnostics_generic_struct_local_annotation_missing_args_schema_matches_golden",
+        "emit_json_diagnostics_generic_enum_local_annotation_arity_schema_matches_golden",
+        "emit_json_diagnostics_generic_enum_local_annotation_missing_args_schema_matches_golden",
+        "emit_json_diagnostics_generic_struct_constructor_arity_schema_matches_golden",
+        "emit_json_diagnostics_generic_struct_constructor_missing_args_schema_matches_golden",
+        "emit_json_diagnostics_nongeneric_struct_constructor_type_args_schema_matches_golden",
+        "emit_json_diagnostics_generic_enum_constructor_arity_schema_matches_golden",
+        "emit_json_diagnostics_generic_enum_constructor_missing_args_schema_matches_golden",
+        "emit_json_diagnostics_nongeneric_enum_constructor_type_args_schema_matches_golden",
+        "emit_json_diagnostics_generic_struct_annotation_arity_schema_matches_golden",
+        "emit_json_diagnostics_nongeneric_struct_annotation_type_args_schema_matches_golden",
+        "emit_json_diagnostics_nongeneric_enum_annotation_type_args_schema_matches_golden",
+        "emit_json_diagnostics_generic_enum_annotation_arity_schema_matches_golden",
+        "emit_json_diagnostics_generic_struct_annotation_missing_args_schema_matches_golden",
+        "emit_json_diagnostics_generic_enum_annotation_missing_args_schema_matches_golden",
+        "emit_json_diagnostics_generic_result_method_arity_schema_matches_golden",
+        "emit_json_diagnostics_generic_result_method_bound_schema_matches_golden",
+        "emit_json_diagnostics_generic_result_method_inference_schema_matches_golden",
+        "emit_json_diagnostics_generic_behavior_overlap_schema_matches_golden",
+        "emit_json_typed_rejects_hand_authored_json_before_checked_ir_override",
+        "emit_json_typed_generic_method_schema_matches_golden",
+        "emit_json_typed_generic_type_impl_methods_schema_matches_golden",
+        "emit_json_typed_generic_self_method_schema_matches_golden",
+        "emit_json_typed_generic_method_worklist_schema_matches_golden",
+        "emit_json_typed_generic_ufc_function_schema_matches_golden",
+        "emit_json_typed_generic_option_schema_matches_golden",
+        "emit_json_typed_generic_result_schema_matches_golden",
+        "emit_json_typed_generic_result_method_schema_matches_golden",
+        "emit_json_typed_nested_generic_result_schema_matches_golden",
+        "emit_json_typed_generic_behavior_association_schema_matches_golden",
+        "emit_json_typed_generic_behavior_bound_ufcs_schema_matches_golden",
+        "Point.encode__Json_Point",
+        "emit_json_build_graph_rejects_hand_authored_json_before_graph_override",
+        "schema_version: 0",
+        "emit_json_hir_outputs_checked_declaration_graph",
+        "emit_json_hir_outputs_enum_function_and_global_declarations",
+        "emit_json_hir_declaration_schema_matches_golden",
+        "emit_json_hir_generic_method_worklist_schema_matches_golden",
+        "emit_json_hir_generic_ufc_function_schema_matches_golden",
+        "emit_json_hir_generic_option_schema_matches_golden",
+        "emit_json_hir_generic_result_schema_matches_golden",
+        "emit_json_hir_nested_generic_result_schema_matches_golden",
+        "emit_json_hir_generic_behavior_association_schema_matches_golden",
+        "emit_json_hir_generic_behavior_bound_ufcs_schema_matches_golden",
+        "emit_json_hir_rejects_hand_authored_json_before_ir_override",
+        "emit_json_mir_outputs_checked_minimal_function_graph",
+        "emit_json_mir_minimal_function_schema_matches_golden",
+        "emit_json_mir_outputs_match_arm_schema",
+        "emit_json_mir_match_schema_matches_golden",
+        "emit_json_mir_generic_method_worklist_schema_matches_golden",
+        "emit_json_mir_generic_ufc_function_schema_matches_golden",
+        "emit_json_mir_generic_option_schema_matches_golden",
+        "emit_json_mir_generic_result_schema_matches_golden",
+        "emit_json_mir_nested_generic_result_schema_matches_golden",
+        "emit_json_mir_generic_behavior_association_schema_matches_golden",
+        "emit_json_mir_generic_behavior_bound_ufcs_schema_matches_golden",
+        "emit_json_mir_rejects_hand_authored_json_before_ir_override",
+        "emit_json_layout_outputs_checked_type_layouts",
+        "emit_json_layout_basic_schema_matches_golden",
+        "emit_json_layout_outputs_compound_type_layout_entries",
+        "emit_json_layout_compound_schema_matches_golden",
+        "emit_json_layout_generic_option_schema_matches_golden",
+        "emit_json_layout_generic_result_schema_matches_golden",
+        "emit_json_layout_nested_generic_result_schema_matches_golden",
+        "emit_json_layout_rejects_hand_authored_json_before_layout_override",
+        "emit_json_target_yaml_validates_minimal_target_schema",
+        "emit_json_target_yaml_validates_backend_schema",
+        "emit_json_target_yaml_validates_c_backend_flags",
+        "emit_json_target_yaml_backend_schema_matches_golden",
+        "emit_json_target_yaml_rejects_empty_c_backend_flags",
+        "emit_json_target_yaml_rejects_layout_overrides",
+        "emit_json_target_yaml_rejects_unsupported_backend_codegen",
+        "emit_json_build_graph_project_schema_matches_golden",
+        "emit_json_build_graph_host_effect_schema_matches_golden",
+        "emit_json_build_graph_target_metadata_schema_matches_golden",
+    ] {
+        assert!(
+            spec.contains(required),
+            "docs/V1_SPEC.md is missing JSON boundary or golden coverage: {required}"
+        );
+    }
+}
